# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "screen-service API",
        "title": "screen-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/screen-service",
    "paths": {
...
        "/topology": {
            "get": {
                "description": "分页查询topo实例",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topology"
                ],
                "summary": "分页查询topo实例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "current page",
                        "name": "_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "_page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例的数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/entity.Topology"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "保存topology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topology"
                ],
                "summary": "保存topology",
                "parameters": [
                    {
                        "description": "topology全部信息",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Topology"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Topology"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/topology-query": {
            "get": {
                "description": "根据条件查询, 可设置 order, 以及查询条件等，例如 status=1, name=$like.%25%25CAM%25%25 等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topology"
                ],
                "summary": "根据条件查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order",
                        "name": "_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "select",
                        "name": "_select",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_time",
                        "name": "created_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_by",
                        "name": "updated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_time",
                        "name": "updated_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "group_id",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "file_data",
                        "name": "file_data",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "key_name",
                        "name": "key_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "parent_id",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "remark",
                        "name": "remark",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例的数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Topology"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/topology/{id}": {
            "delete": {
                "description": "删除topology",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topology"
                ],
                "summary": "删除topology",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实例id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Topology"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Page": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Topology": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建人",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "fileData": {
                    "description": "文件内容",
                    "type": "string"
                },
                "groupId": {
                    "description": "分组id",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "keyName": {
                    "description": "标识",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父id",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "type": {
                    "description": "类型",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新人",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "model.Topology": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "创建人",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "file_data": {
                    "description": "文件内容",
                    "type": "string"
                },
                "group_id": {
                    "description": "分组id",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "key_name": {
                    "description": "标识",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父id",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "type": {
                    "description": "类型",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新人",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        }
    }
}